Customer formatting and one class per file
