+++ code/Customer/AccountPage.php
@@ -7,11 +7,6 @@ use SilverStripe\ORM\DB;
 use SilverStripe\Security\Member;
 use SilverStripe\Security\Permission;
 use SilverStripe\Control\Director;
-use SilverStripe\View\Requirements;
-use SwipeStripe\Core\code\Order\Order;
-use SilverStripe\Core\Convert;
-use SilverStripe\Control\Session;
-use SwipeStripe\Core\code\Form\RepayForm;
 
 /**
  * An account page which displays the order history for any given {@link Member} and displays an individual {@link Order}.
@@ -25,7 +20,6 @@ use SwipeStripe\Core\code\Form\RepayForm;
  */
 class AccountPage extends Page
 {
-
     /**
      * Automatically create an AccountPage if one is not found
      * on the site at the time the database is built (dev/build).
@@ -46,7 +40,7 @@ class AccountPage extends Page
             DB::alteration_message('Account page \'Account\' created', 'created');
         }
     }
-    
+
     /**
      * Prevent CMS users from creating another account page.
      *
@@ -61,7 +55,7 @@ class AccountPage extends Page
         }
         return false;
     }
-    
+
     /**
      * Prevent CMS users from deleting the account page.
      *
@@ -83,7 +77,7 @@ class AccountPage extends Page
             parent::delete();
         }
     }
-    
+
     /**
      * Prevent CMS users from unpublishing the account page.
      *
@@ -99,7 +93,7 @@ class AccountPage extends Page
         }
         return false;
     }
-    
+
     /**
      * To remove the unpublish button from the CMS, as this page must always be published
      *
@@ -113,7 +107,7 @@ class AccountPage extends Page
         $actions->removeByName('action_unpublish');
         return $actions;
     }
-    
+
     /**
      * Remove page type dropdown to prevent users from changing page type.
      *
@@ -125,134 +119,5 @@ class AccountPage extends Page
         $fields = parent::getCMSFields();
         $fields->removeByName('ClassName');
         return $fields;
-    }
-}
-
-/**
- * Display the account page with listing of previous orders, and display an individual order.
- *
- * @author Frank Mullenger <frankmullenger@gmail.com>
- * @copyright Copyright (c) 2011, Frank Mullenger
- * @package swipestripe
- * @subpackage customer
- */
-class AccountPage_Controller extends PageController
-{
-    
-    /**
-     * Allowed actions that can be invoked.
-     *
-     * @var Array Set of actions
-     */
-    private static $allowed_actions = array(
-        'index',
-        'order',
-        'repay',
-        'RepayForm'
-    );
-    
-    public function init()
-    {
-        parent::init();
-
-        if (!Permission::check('VIEW_ORDER')) {
-            return $this->redirect(Director::absoluteBaseURL() . 'Security/login?BackURL=' . urlencode($this->getRequest()->getVar('url')));
-        }
-    }
-    
-    /**
-     * Check access permissions for account page and return content for displaying the
-     * default page.
-     *
-     * @return Array Content data for displaying the page.
-     */
-    public function index()
-    {
-        Requirements::css('swipestripe/css/Shop.css');
-
-        return array(
-            'Content' => $this->Content,
-            'Form' => $this->Form,
-            'Orders' => Order::get()
-                ->where("MemberID = " . Convert::raw2sql(Member::currentUserID()))
-                ->sort('Created DESC'),
-            'Customer' => Customer::currentUser()
-        );
-    }
-
-    /**
-     * Return the {@link Order} details for the current Order ID that we're viewing (ID parameter in URL).
-     *
-     * @return Array Content for displaying the page
-     */
-    public function order($request)
-    {
-        Requirements::css('swipestripe/css/Shop.css');
-
-        if ($orderID = $request->param('ID')) {
-            $member = Customer::currentUser();
-            $order = Order::get()
-                ->where("\"Order\".\"ID\" = " . Convert::raw2sql($orderID))
-                ->First();
-
-            if (!$order || !$order->exists()) {
-                return $this->httpError(403, _t('AccountPage.NO_ORDER_EXISTS', 'Order does not exist.'));
-            }
-
-            if (!$order->canView($member)) {
-                return $this->httpError(403, _t('AccountPage.CANNOT_VIEW_ORDER', 'You cannot view orders that do not belong to you.'));
-            }
-
-            return array(
-                'Order' => $order
-            );
-        } else {
-            return $this->httpError(403, _t('AccountPage.NO_ORDER_EXISTS', 'Order does not exist.'));
-        }
-    }
-    
-    public function repay($request)
-    {
-        Requirements::css('swipestripe/css/Shop.css');
-
-        if ($orderID = $request->param('ID')) {
-            $member = Customer::currentUser();
-            $order = Order::get()
-                ->where("\"Order\".\"ID\" = " . Convert::raw2sql($orderID))
-                ->First();
-
-            if (!$order || !$order->exists()) {
-                return $this->httpError(403, _t('AccountPage.NO_ORDER_EXISTS', 'Order does not exist.'));
-            }
-
-            if (!$order->canView($member)) {
-                return $this->httpError(403, _t('AccountPage.CANNOT_VIEW_ORDER', 'You cannot view orders that do not belong to you.'));
-            }
-            
-            Session::set('Repay', array(
-                'OrderID' => $order->ID
-            ));
-            Session::save();
-            
-            return array(
-                'Order' => $order,
-                'RepayForm' => $this->RepayForm()
-            );
-        } else {
-            return $this->httpError(403, _t('AccountPage.NO_ORDER_EXISTS', 'Order does not exist.'));
-        }
-    }
-
-    public function RepayForm()
-    {
-        $form = RepayForm::create(
-            $this,
-            RepayForm::class
-        )->disableSecurityToken();
-
-        //Populate fields the first time form is loaded
-        $form->populateFields();
-
-        return $form;
     }
 }
